requirements.txt
# Django et extensions
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
django-environ==0.12.0
drf-gis==1.1
drf-spectacular==0.28.0
python-dotenv==1.1.0

# Base de données
psycopg2-binary==2.9.9

# Tâches asynchrones
celery==5.5.3
redis==6.4.0

# APIs externes et données cartographiques
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
folium==0.15.1
matplotlib==3.8.2
pillow==10.1.0
numpy==1.24.4
pandas==2.1.4

# Géospatial (installé via conda)
# GDAL installé via conda-forge
